requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bleach==6.2.0
colour==0.1.5
ddt==1.7.2
Django==5.1.6
django-cors-headers==4.7.0
django-markdownify==0.9.5
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
Markdown==3.7
markdownify==1.1.0
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.12.2
webencodings==0.5.1
python-dotenv~=1.1.1